requirements.txt
flask
requests
python-dotenv
twilio
joblib
pandas
numpy
scikit-learn
networkx
osmnx
gunicorn
